+++ templates/generate.js
@@ -1,6 +1,0 @@
-const config = require(process.argv[2]);
-const versionconfig = require("../package.json");
-
-config.dependencies["@mindconnect/mindconnect-nodejs"] = `file:mindconnect-mindconnect-nodejs-${versionconfig.version}.tgz`;
-
-console.log(JSON.stringify(config, null, 2));